pyproject.toml
[tool.poetry]
name = "asyncapi_demo"
version = "0.1.0"
description = ""
authors = ["TATA <evadcmd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.2"}
aioredis = "^1.3.1"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyre-check = "^0.9.0"
black = "^21.4b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
